// src/lib/types.ts
export interface Question {
  id: string | number;
  questionText: string;
  imageUrl?: string;
  options: string[];
  correctAnswer: string;
}

export interface Exam {
  id: string; // Document ID generated by Firestore
  examTitle: string;
  examDescription: string;
  questions: Question[];
  creatorEmail: string;
  createdAt: number;
}

export interface ExamResult {
  id?: string;
  examId: string;
  examTitle: string;
  studentEmail: string;
  score: number;
  totalQuestions: number;
  submittedAt: number;
  studentAnswers?: Record<string, string>;
}